% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMQFuture-class.R
\name{ClusterMQFuture}
\alias{ClusterMQFuture}
\alias{run.ClusterMQFuture}
\title{A clustermq future is a future whose value will be resolved via clustermq}
\usage{
ClusterMQFuture(expr = NULL, envir = parent.frame(),
  substitute = TRUE, globals = TRUE, packages = NULL, label = NULL,
  workers = Inf, ...)
}
\arguments{
\item{expr}{The \R expression to be evaluated.}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be \code{substitute()}:d
or not.}

\item{globals}{(optional) a logical, a character vector, a named list, or
a \link[globals:Globals]{globals::Globals} object.  If \code{TRUE}, globals are identified by code
inspection based on \code{expr} and \code{tweak} searching from environment
\code{envir}.  If \code{FALSE}, no globals are used.  If a character vector, then
globals are identified by lookup based their names \code{globals} searching
from environment \code{envir}.  If a named list or a Globals object, the
globals are used as is.}

\item{label}{(optional) Label of the future.}

\item{workers}{(optional) The maximum number of workers the clustermq
backend may use at any time.}

\item{\ldots}{Additional arguments passed to \code{\link[future:MultiprocessFuture]{future::MultiprocessFuture()}}.}
}
\value{
A ClusterMQFuture object
}
\description{
A clustermq future is a future whose value will be resolved via clustermq
}
\keyword{internal}
