% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMQFuture-class.R
\name{print.ClusterMQFuture}
\alias{print.ClusterMQFuture}
\title{Prints a clustermq future}
\usage{
\method{print}{ClusterMQFuture}(x, ...)
}
\arguments{
\item{x}{An ClusterMQFuture object}

\item{\ldots}{Not used.}
}
\description{
Prints a clustermq future
}
\keyword{internal}
